FEAT: auto click after get_action response
